skrypt do uruchamiania przeglądarki

+++ src/main/java/com/crud/tasks/service/DbService.java
@@ -1,7 +1,6 @@
 package com.crud.tasks.service;
 
 import com.crud.tasks.domain.Task;
-import com.crud.tasks.domain.TaskDto;
 import com.crud.tasks.repository.TaskRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,6 +27,7 @@ public class DbService {
     }
 
     public void deleteTask(final Long taskId) {
-        repository.deleteById(taskId);
+        if (repository.existsById(taskId))
+            repository.deleteById(taskId);
     }
 }
